History: Unlock YAML export format
